Move tests to test_auth.py
